[#36] applied differently fix
I overlooked the 'old' solution in `setup-ropsi.sh` that was copying the
cantera module from home user directory to
`/usr/bin/pythonX.Y/dist-packages` folder. I extended `setup-ropsi.sh`
to deal with Python2.7 and Python3.4 folders.

This required some changes in comments to `setup-cantera.sh` that are
applied in this commit.